Rename .h file extension to .hpp
Prepare for spearating the C and the C++ interface.
