+++ SlashGaming-Diablo-II-API/include/game_constant.hpp
@@ -35,7 +35,7 @@
  *  work.
  */
 
-#ifndef SGD2MAPI_GAME_CONSTANT_H_
-#define SGD2MAPI_GAME_CONSTANT_H_
+#ifndef SGD2MAPI_GAME_CONSTANT_HPP_
+#define SGD2MAPI_GAME_CONSTANT_HPP_
 
-#endif // SGD2MAPI_GAME_CONSTANT_H_
+#endif // SGD2MAPI_GAME_CONSTANT_HPP_
