+++ SlashGaming-Diablo-II-API/include/game_patch/game_buffer_patch.hpp
@@ -35,14 +35,14 @@
  *  work.
  */
 
-#ifndef SGD2MAPI_GAME_PATCH_GAME_BUFFER_PATCH_H_
-#define SGD2MAPI_GAME_PATCH_GAME_BUFFER_PATCH_H_
+#ifndef SGD2MAPI_GAME_PATCH_GAME_BUFFER_PATCH_HPP_
+#define SGD2MAPI_GAME_PATCH_GAME_BUFFER_PATCH_HPP_
 
 #include <stdint.h>
 #include <stdlib.h>
 
-#include "../game_address.h"
-#include "game_patch_base.h"
+#include "../game_address.hpp"
+#include "game_patch_base.hpp"
 
 #ifdef __cplusplus
 #include <cstdint>
@@ -266,4 +266,4 @@ SGD2MAPI_GameBufferPatch_Remove(
 #endif // !defined(__cplusplus) || defined(SGD2MAPI_DLLEXPORT)
 
 #undef DLLEXPORT
-#endif // SGD2MAPI_GAME_PATCH_GAME_BUFFER_PATCH_H_
+#endif // SGD2MAPI_GAME_PATCH_GAME_BUFFER_PATCH_HPP_
