+++ SlashGaming-Diablo-II-API/include/sgd2mapi.hpp
@@ -35,96 +35,18 @@
  *  work.
  */
 
-#ifndef SGD2MAPI_GAME_LIBRARY_H_
-#define SGD2MAPI_GAME_LIBRARY_H_
+#ifndef SGD2MAPI_SGD2MAPI_HPP_
+#define SGD2MAPI_SGD2MAPI_HPP_
 
-#include <cstdint>
-#include <filesystem>
+#include "default_game_library.hpp"
+#include "game_address.hpp"
+#include "game_address_locator.hpp"
+#include "game_bool.hpp"
+#include "game_constant.hpp"
+#include "game_data.hpp"
+#include "game_func.hpp"
+#include "game_patch.hpp"
+#include "game_struct.hpp"
+#include "game_version.hpp"
 
-#include "../include/default_game_library.h"
-
-namespace sgd2mapi {
-
-/**
- * A class that corresponds to a library used by the game.
- */
-class GameLibrary {
- public:
-  /**
-   * Creates a new instance of a GameLibrary using the default library ID.
-   */
-  explicit GameLibrary(
-      enum DefaultLibrary library
-  );
-
-  /**
-   * Creates a new instance of a GameLibrary using the library path.
-   */
-  explicit GameLibrary(
-      const std::filesystem::path& library_path
-  );
-
-  /**
-   * Creates a new instance of a GameLibrary using the library path.
-   */
-  explicit GameLibrary(
-      std::filesystem::path&& library_path
-  );
-
-  GameLibrary(
-      const GameLibrary& rhs
-  );
-
-  GameLibrary(
-      GameLibrary&& rhs
-  );
-
-  virtual
-  ~GameLibrary(
-      void
-  );
-
-  GameLibrary&
-  operator=(
-      const GameLibrary& rhs
-  );
-
-  GameLibrary&
-  operator=(
-      GameLibrary&& rhs
-  );
-
-  /**
-   * Returns the base address value of this GameLibrary.
-   */
-  std::intptr_t
-  base_address(
-      void
-  ) const noexcept;
-
-  /**
-   * Returns the library path of this GameLibrary.
-   */
-  const std::filesystem::path&
-  library_path(
-      void
-  ) const noexcept;
-
- private:
-  std::filesystem::path library_path_;
-  std::intptr_t base_address_;
-};
-
-const GameLibrary&
-GetGameLibrary(
-    enum DefaultLibrary library
-);
-
-const GameLibrary&
-GetGameLibrary(
-    const std::filesystem::path& library_path
-);
-
-} // namespace sgd2mapi
-
-#endif // SGD2MAPI_GAME_LIBRARY_H_
+#endif // SGD2MAPI_SGD2MAPI_HPP_
