+++ SlashGaming-Diablo-II-API/src/config_parser.hpp
@@ -35,7 +35,20 @@
  *  work.
  */
 
-#ifndef SGD2MAPI_GAME_STRUCT_H_
-#define SGD2MAPI_GAME_STRUCT_H_
+#ifndef SGD2MAPI_CONFIG_PARSER_HPP_
+#define SGD2MAPI_CONFIG_PARSER_HPP_
 
-#endif // SGD2MAPI_GAME_STRUCT_H_
+#include <filesystem>
+#include <string>
+#include <string_view>
+
+namespace sgd2mapi {
+
+const std::filesystem::path&
+GetAddressTableDirectoryPath(
+    void
+);
+
+} // namespace sgd2mapi
+
+#endif // SGD2MAPI_CONFIG_PARSER_HPP_
