+++ SlashGaming-Diablo-II-API/src/default_game_library.cc
@@ -35,7 +35,7 @@
  *  work.
  */
 
-#include "../include/default_game_library.h"
+#include "../include/default_game_library.hpp"
 
 #include <windows.h>
 #include <cstdint>
@@ -48,7 +48,7 @@
 
 #include <fmt/format.h>
 #include <fmt/printf.h>
-#include "../include/game_version.h"
+#include "../include/game_version.hpp"
 
 namespace sgd2mapi {
 namespace {
