+++ SlashGaming-Diablo-II-API/src/game_address.cc
@@ -35,7 +35,7 @@
  *  work.
  */
 
-#include "../include/game_address.h"
+#include "../include/game_address.hpp"
 
 #include <windows.h>
 #include <filesystem>
@@ -46,10 +46,10 @@
 
 #include <fmt/format.h>
 #include <fmt/printf.h>
-#include "../include/game_address_locator.h"
-#include "../include/default_game_library.h"
-#include "game_library.h"
-#include "../include/game_version.h"
+#include "../include/game_address_locator.hpp"
+#include "../include/default_game_library.hpp"
+#include "game_library.hpp"
+#include "../include/game_version.hpp"
 
 namespace sgd2mapi {
 namespace {
