+++ SlashGaming-Diablo-II-API/src/game_address_locator/game_address_locator_interface.cc
@@ -35,7 +35,7 @@
  *  work.
  */
 
-#include "../../include/game_address_locator/game_address_locator_interface.h"
+#include "../../include/game_address_locator/game_address_locator_interface.hpp"
 
 namespace sgd2mapi {
